NO-ISSUE: Move repotester to tasks package
